+++ DogBarberShop_service/src/main/java/cz/fi/muni/pa165/dogbarber/facade/EmployeeFacadeImpl.java
@@ -4,6 +4,7 @@ import cz.fi.muni.pa165.dogbarber.entity.Employee;
 import cz.fi.muni.pa165.dogbarber.dto.EmployeeDTO;
 import cz.fi.muni.pa165.dogbarber.dto.ServiceDTO;
 import cz.fi.muni.pa165.dogbarber.entity.Service;
+import cz.fi.muni.pa165.dogbarber.exception.DogBarberException;
 import cz.fi.muni.pa165.dogbarber.service.BeanMappingService;
 import cz.fi.muni.pa165.dogbarber.service.EmployeeService;
 import java.util.Collection;
@@ -33,6 +34,10 @@ public class EmployeeFacadeImpl implements EmployeeFacade {
 
     @Override
     public void deleteEmployee(EmployeeDTO emp) {
+         if(emp==null){
+            throw new DogBarberException("Employee does not exist!");
+        }
+        
         employeeService.removeEmployee(beanMappingService.mapTo(emp, Employee.class));
     }
     
@@ -53,8 +58,7 @@ public class EmployeeFacadeImpl implements EmployeeFacade {
         Employee emp = employeeService.findEmployeeById(userId);
         
          if(emp==null){
-            return null;
-        
+            throw new DogBarberException("Employee does not exist!");
         }
             return beanMappingService.mapTo(emp, EmployeeDTO.class);
     
@@ -63,6 +67,10 @@ public class EmployeeFacadeImpl implements EmployeeFacade {
     @Override
     public List<EmployeeDTO> findEmployeeByName(String name) {
         List<Employee> users = employeeService.findEmployeeByName(name);
+        
+         if(users==null){
+            throw new DogBarberException("Employees does not exist!");
+        }
         return beanMappingService.mapTo(users, EmployeeDTO.class);
     }
 
@@ -73,25 +81,42 @@ public class EmployeeFacadeImpl implements EmployeeFacade {
     
     @Override
     public EmployeeDTO update(EmployeeDTO e) {
+         if(e==null){
+            throw new DogBarberException("Employee does not exist!");
+        }
+        
         return beanMappingService.mapTo(employeeService.update(beanMappingService.mapTo(e, Employee.class)),EmployeeDTO.class);
     }
     
     @Override
     public void addService(EmployeeDTO emp, ServiceDTO s) {
+         if(emp==null){
+            throw new DogBarberException("Employee does not exist!");
+        }
+          if(s==null){
+            throw new DogBarberException("Service does not exist!");
+        }
         employeeService.addService(beanMappingService.mapTo(emp,Employee.class), beanMappingService.mapTo(s, Service.class));
     }
     
     @Override
     public void removeService(EmployeeDTO emp, ServiceDTO s) {
+         if(emp==null){
+            throw new DogBarberException("Employee does not exist!");
+        }
+          if(s==null){
+            throw new DogBarberException("Service does not exist!");
+        }
         employeeService.removeService(beanMappingService.mapTo(emp,Employee.class), beanMappingService.mapTo(s, Service.class));
     }
 
     @Override
     public EmployeeDTO findEmployeeByEmail(String email) {
+        
         Employee emp = employeeService.findEmployeeByEmail(email);
         
         if(emp==null){
-            return null;
+            throw new DogBarberException("Employee does not exist!");
         
         }
             return beanMappingService.mapTo(emp, EmployeeDTO.class);
