+++ DogBarberShop_service/src/main/java/cz/fi/muni/pa165/dogbarber/facade/ServiceFacadeImpl.java
@@ -4,6 +4,7 @@ package cz.fi.muni.pa165.dogbarber.facade;
 import cz.fi.muni.pa165.dogbarber.dto.ServiceChangeNameDTO;
 import cz.fi.muni.pa165.dogbarber.dto.ServiceDTO;
 import cz.fi.muni.pa165.dogbarber.entity.Service;
+import cz.fi.muni.pa165.dogbarber.exception.DogBarberException;
 import cz.fi.muni.pa165.dogbarber.service.BeanMappingService;
 import cz.fi.muni.pa165.dogbarber.service.EmployeeService;
 import cz.fi.muni.pa165.dogbarber.service.ServiceService;
@@ -31,17 +32,37 @@ public class ServiceFacadeImpl implements ServiceFacade {
     
     @Override
     public void changeServiceName(ServiceChangeNameDTO newName){
+        if(newName==null){
+            throw new DogBarberException("New name cannot be null");
+        }
+        
         service.changeServiceName(service.findById(newName.getServiceId()), newName.getServiceName());
     
     }
     
     @Override
     public void addEmployee(Long serviceId, Long employeeId ){
+        if(service.findById(serviceId)==null){
+            throw new DogBarberException("Service does not exist!");
+        }
+        
+        if(employeeService.findEmployeeById(employeeId)==null){
+            throw new DogBarberException("Employee does not exist!");
+        }
+        
     service.addEmployee(service.findById(serviceId),employeeService.findEmployeeById(employeeId) );
     }
     
     @Override
     public void deleteEmployee(Long serviceId, Long employeeId){
+        if(service.findById(serviceId)==null){
+            throw new DogBarberException("Service does not exist!");
+        }
+        
+        if(employeeService.findEmployeeById(employeeId)==null){
+            throw new DogBarberException("Employee does not exist!");
+        }
+        
         service.removeEmployee(service.findById(serviceId),employeeService.findEmployeeById(employeeId));
         
     }
@@ -67,12 +88,20 @@ public class ServiceFacadeImpl implements ServiceFacade {
     
     @Override
     public ServiceDTO getServiceById(Long id){
+        
+        if(service.findById(id)==null){
+            throw new DogBarberException("Service does not exist!");
+        }
       
         return beanMappingService.mapTo(service.findById(id), ServiceDTO.class);
     }
     
     @Override
     public Long createService(ServiceDTO s){
+        if(s==null){
+            throw new IllegalArgumentException("Service cannot be null!");
+        }
+        
     
         cz.fi.muni.pa165.dogbarber.entity.Service serv = new cz.fi.muni.pa165.dogbarber.entity.Service();
         serv.setServiceName(s.getServiceName());
@@ -88,6 +117,10 @@ public class ServiceFacadeImpl implements ServiceFacade {
     
     @Override
     public void deleteService(Long serviceId){
+        
+        if(service.findById(serviceId)==null){
+            throw new DogBarberException("Service does not exist!");
+        }
     
         service.remove(service.findById(serviceId));
     }
@@ -104,6 +137,10 @@ public class ServiceFacadeImpl implements ServiceFacade {
         serviceEntity.setServiceName(s.getServiceName());
         service.update(serviceEntity);
     */
+        if(s==null){
+            throw new DogBarberException("Service does not exist!");
+        }
+        
         service.update(beanMappingService.mapTo(s, cz.fi.muni.pa165.dogbarber.entity.Service.class));
     }
